HW#15 Created interface for TrafficGenerator
